requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.0
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
PyJWT==2.10.1
qrcode[pil]==8.1
bcrypt==4.1.2
python-dotenv==1.0.1
